[CHORE] SelectList Component 이름 변경
